Reverted images.go back. This pr is purely firecracker
